# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Intelligence & Question Extraction Service",
    "description": "\n    ## Pragati Bharati Document Intelligence & Question Extraction Service\n    \n    A production-oriented, scalable microservice to ingest examination papers and question banks\n    in PDF (digital and scanned) and Image formats, converting them into structured, machine-readable\n    question sets.\n    \n    ### Key Features:\n    * **Asynchronous Document Ingestion**: Upload large documents without blocking clients.\n    * **Automated Question Segmentation**: Detect question numbers, bodies, and options.\n    * **Cross-Page Stitching**: Intelligently handles questions continuing across page breaks.\n    * **Answer Key Matching**: Pairs answer keys from inline sections or standalone documents.\n    * **Confidence & Review Pipeline**: Flags uncertain, partial, or low-quality extractions.\n    * **Security & Auth**: JWT-based bearer authentication and user authorization isolation.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "Upload a PDF or Image asynchronously.\nReturns 202 Accepted immediately with document_id and tracking status.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDetailResponse"
                  },
                  "title": "Response List Documents Api V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document Details",
        "operationId": "get_document_details_api_v1_documents__document_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Check Document Status",
        "operationId": "check_document_status_api_v1_documents__document_id__status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/page-preview/{page_num}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Preview Page",
        "description": "Render a page as PNG image for inspection or reviewer dashboard",
        "operationId": "preview_page_api_v1_documents__document_id__page_preview__page_num__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "page_num",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Num"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/relate": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Relate Documents",
        "description": "Associate Answer Key document with Question Paper document and trigger cross-matching.",
        "operationId": "relate_documents_api_v1_documents__document_id__relate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRelateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRelateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/reprocess": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Reprocess Document",
        "operationId": "reprocess_document_api_v1_documents__document_id__reprocess_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "document_role": {
            "$ref": "#/components/schemas/DocumentRole",
            "default": "QUESTION_PAPER"
          },
          "target_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Document Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "DocumentCreateResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "original_filename",
          "status",
          "message",
          "task_id",
          "created_at"
        ],
        "title": "DocumentCreateResponse"
      },
      "DocumentDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "document_role": {
            "$ref": "#/components/schemas/DocumentRole"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "ocr_method_used": {
            "type": "string",
            "title": "Ocr Method Used"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "confident_questions": {
            "type": "integer",
            "title": "Confident Questions"
          },
          "review_required_questions": {
            "type": "integer",
            "title": "Review Required Questions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "original_filename",
          "file_size_bytes",
          "mime_type",
          "document_type",
          "document_role",
          "status",
          "progress_percent",
          "page_count",
          "ocr_method_used",
          "total_questions",
          "confident_questions",
          "review_required_questions",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentDetailResponse"
      },
      "DocumentRelateRequest": {
        "properties": {
          "target_document_id": {
            "type": "string",
            "title": "Target Document Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationType",
            "default": "ANSWER_KEY_FOR"
          }
        },
        "type": "object",
        "required": [
          "target_document_id"
        ],
        "title": "DocumentRelateRequest"
      },
      "DocumentRelateResponse": {
        "properties": {
          "relation_id": {
            "type": "string",
            "title": "Relation Id"
          },
          "source_document_id": {
            "type": "string",
            "title": "Source Document Id"
          },
          "target_document_id": {
            "type": "string",
            "title": "Target Document Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationType"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "relation_id",
          "source_document_id",
          "target_document_id",
          "relationship_type",
          "message",
          "created_at"
        ],
        "title": "DocumentRelateResponse"
      },
      "DocumentRole": {
        "type": "string",
        "enum": [
          "QUESTION_PAPER",
          "ANSWER_KEY",
          "COMBINED",
          "SUPPLEMENTARY"
        ],
        "title": "DocumentRole"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "DocumentStatus"
      },
      "DocumentStatusResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "confident_questions": {
            "type": "integer",
            "title": "Confident Questions"
          },
          "review_required_questions": {
            "type": "integer",
            "title": "Review Required Questions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "original_filename",
          "status",
          "progress_percent",
          "page_count",
          "total_questions",
          "confident_questions",
          "review_required_questions",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentStatusResponse"
      },
      "DocumentType": {
        "type": "string",
        "enum": [
          "PDF",
          "IMAGE"
        ],
        "title": "DocumentType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationType": {
        "type": "string",
        "enum": [
          "ANSWER_KEY_FOR",
          "SUPPLEMENT_FOR",
          "RELATED_TO"
        ],
        "title": "RelationType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}